+++ lib/ayeaye/connect.py
@@ -3,19 +3,16 @@ from enum import Enum
 import glob
 
 from ayeaye.connectors import connector_factory
-from ayeaye.connectors.models_connector import ModelsConnector
 from ayeaye.connectors.multi_connector import MultiConnector
 from ayeaye.connectors.placeholder import PlaceholderDataConnector
 
 
 class Connect:
     """
-    Connect to a dataset or modelset.
+    Connect to a dataset.
 
     A dataset is a grouping of related data. What makes data 'related' will be down to the
     subject-domain. Datasets have a type - it could be a JSON document, a CSV file or a database.
-
-    A modelset is one or more :class:`Model`(s).
 
     The main responsibility of :class:`Connect` is to provide a concrete subclass of
     :class:`DataConnector` or :class:`ModelConnector`. :class:`DataConnector` in turn provides
@@ -45,7 +42,7 @@ class Connect:
     For secrets management @see :class:`ConnectorResolver`.
     """
 
-    mutually_exclusive_selectors = ["ref", "engine_url", "models"]
+    mutually_exclusive_selectors = ["ref", "engine_url"]
 
     class ConnectBind(Enum):
         MODEL = "MODEL"
@@ -77,7 +74,8 @@ class Connect:
         a = [self.relayed_kwargs.get(s) is not None for s in self.mutually_exclusive_selectors]
         mandatory_args_count = sum(a)
         if mandatory_args_count > 1:
-            raise ValueError("The kwargs ref, engine_url and models are mutually exclusive.")
+            mut_ex_field = " and ".join(self.mutually_exclusive_selectors)
+            raise ValueError(f"For kwargs, {mut_ex_field} values are mutually exclusive.")
 
         self.ref = self.relayed_kwargs.pop("ref", None)
         self._standalone_connection = None  # see :method:`data`
@@ -114,7 +112,7 @@ class Connect:
     def connect_id(self):
         """
         Create an identity reference which is used when examining if separate Connect instances
-        are actually referring to the same dataset/models.
+        are actually referring to the same datasets.
 
         @return: (str)
         """
@@ -191,11 +189,7 @@ class Connect:
         if callable(engine_url):
             engine_url = engine_url()
 
-        if "models" in self.relayed_kwargs:
-            # could be a callable but shouldn't be instantiated yet, ModelsConnector does that
-            connector_cls = ModelsConnector
-
-        elif engine_url is None:
+        if engine_url is None:
             # engine_url not yet available
             connector_cls = PlaceholderDataConnector
 
@@ -230,7 +224,7 @@ class Connect:
                 # this might have been a callable above
                 detached_kwargs[k] = copy.deepcopy(engine_url)
 
-            elif callable(v) and k != "models":
+            elif callable(v):
 
                 if k in connector_cls.preserve_callables:
                     # reference to the callable is preserved for the `connector_cls` to call. This is