+++ lib/ayeaye/connectors/__init__.py
@@ -14,19 +14,28 @@ from .parquet_connector import ParquetConnector
 from .sqlalchemy_database import SqlAlchemyDatabaseConnector
 from .uncooked_connector import UncookedConnector
 
-class ConnectorPluginsRegistry:
 
+class ConnectorPluginsRegistry:
     def __init__(self):
-        self.registered_connectors = [] # list of classes, not instances - publicly readable
+        self.registered_connectors = []  # list of classes, not instances - publicly readable
         self.reset()
 
     def reset(self):
         "set registered connectors to just those that are built in"
-        self.registered_connectors = [BigQueryConnector, CsvConnector, FlowerPotConnector,
-                                      FakeDataConnector, KafkaConnector, ParquetConnector,
-                                      TsvConnector, SqlAlchemyDatabaseConnector, JsonConnector,
-                                      NdjsonConnector, ElasticsearchConnector, UncookedConnector,
-                                      ]
+        self.registered_connectors = [
+            BigQueryConnector,
+            CsvConnector,
+            FlowerPotConnector,
+            FakeDataConnector,
+            KafkaConnector,
+            ParquetConnector,
+            TsvConnector,
+            SqlAlchemyDatabaseConnector,
+            JsonConnector,
+            NdjsonConnector,
+            ElasticsearchConnector,
+            UncookedConnector,
+        ]
 
     def register_connector(self, connector_cls):
         """
@@ -43,13 +52,14 @@ class ConnectorPluginsRegistry:
 # global registry
 connector_registry = ConnectorPluginsRegistry()
 
+
 def connector_factory(engine_url):
     """
     return a subclass of DataConnector
     @param engine_url (str):
     """
     # TODO - these really need to be on demand and possibly outside this project
-    engine_type = engine_url.split('://', 1)[0] + '://'
+    engine_type = engine_url.split("://", 1)[0] + "://"
     for connector_cls in connector_registry.registered_connectors:
         if isinstance(connector_cls.engine_type, list):
             supported_engines = connector_cls.engine_type
@@ -60,4 +70,3 @@ def connector_factory(engine_url):
             return connector_cls
 
     raise NotImplementedError(f"Unknown engine in url:{engine_url}")
-