+++ CHANGELOG.md
@@ -4,6 +4,37 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased - 2022-07-12
+
+### Added
+- ModelCollection is top level public class
+- ModelCollection can now build data provenance graphs for simple model collections - i.e. those that
+ don't need any resolver context or locking info
+- ModelCollection is incomplete and has a failing test
+
+## [0.0.26] - 2022-06-09
+
+### Fixed
+- multiprocess support for ayeaye.PartitionedModel had different behaviour on OSX and Linux because of parent memory after fork() with resolver context
+
+## [0.0.25] - 2022-06-08
+
+### Added
+- manifest_build_context() common pattern - make values from a manifest file available in the resolver context
+- example usage for ayeaye.common_pattern.build_context.manifest_build_context
+
+### Changed
+- ayeaye.connect_resolve supports lists, dict and anything else JSON serialisable. Previously just simple variables.
+
+## [0.0.24] - 2022-05-09
+
+### Removed
+- ModelsConnector - it makes better sense outside the ayeaye.Connect(..) declaration
+
+### Added
+- ModelCollection - the old ModelsConnector slightly repackaged
+- ayeaye.common_pattern.build_context.manifest_build_context and examples/manifest_build_context to demonstrate a simple way to list files used in an ayeaye.Model
+
 ## [0.0.23] - 2022-03-10
 
 ### Added